stop record user activity if statistics disabled, move config operation to app/config.py
