+++ main.py
@@ -7,9 +7,9 @@ import os
 
 
 from seafevents import is_audit_enabled
-from seafevents.utils import write_pidfile
+from seafevents.utils import write_pidfile, get_config
 from seafevents.app.log import LogConfigurator
-from seafevents.app.app import App, get_config
+from seafevents.app.app import App
 from seafevents.app.mq_listener import init_message_handlers
